fix: users api called multiple times
